fix: cast advisory lock key to bigint in SwapLiveReplyPost
Validated against postgres 16 with an ephemeral DB.

+++ internal/store/postgres/raw_event_repository.go
@@ -571,7 +571,7 @@ func (rawEventRepository RawEventRepository) SwapLiveReplyPost(taskRunID string,
 	}
 	defer transaction.Rollback()
 
-	if _, errorValue := transaction.Exec(`SELECT pg_advisory_xact_lock(hashtext($1))`, taskRunID+"\x00"+conversationID); errorValue != nil {
+	if _, errorValue := transaction.Exec(`SELECT pg_advisory_xact_lock(hashtext($1)::bigint)`, taskRunID+"\x00"+conversationID); errorValue != nil {
 		return "", errorValue
 	}
